chore(frontend): fix notification context fast-refresh warning
